Convert class to module
